Rename container to features
